Net::DNS::Simple: optionally disable non-standard extensions
Net::EPP::Simple echos the entire server feature list when it logs
in. Nominet's EPP server rejects this with a "2100 Unimplemented
protocol version" error.

The new stdext option filters out non-standard features, i.e. those
not starting urn:, which is enough to solve the immediate problem
but it isn't a very good fix.